Change to use pointers to commands and events.
As of this version commands and events are recommended to be passed
around as pointers, instead of values as the previous versions did.
Passing as values may still work, but is not tested at the momemnt.
It should not requrie much changes in applications using
Event Horizon, simple pass all commands and events with & before them
or create them as *XXXCommand, see the examples and tests for usage.
There are also some other API changes to method names, mostly with
using "handler" as a more common term.