Add improve Aggregate event handling
ApplyEvent no longer needs to call IncrementVersion, that is now
handled by the aggregate repository upon loading and saving aggregates.
NewEventForAggregate is added to help create events in AggregateBase
which have the correct aggregate and version set. ApplyEvent now returns
an error if the event could not be applied, which aborts any saving or
loading operation.