Fix dynamo event data structure
The previous version would convert the raw event data to a map of map containing all but one null values. Seems like the issue was with how the MarshalMap function handles the data structure. I could not find a clean way to fix this without using the github.com/guregu/dynamo library that already handles this kind of marshalling.

From now on the RawData field contains a map of only the real value (i.e a map of strings and numbers instead of a map of map of mostly null values)